Simplify the passing of sandbox options around and reduce modality.
Split the sandbox option factory function into two - the mode was always
a static arg and never a dynamic variable, so it isn't necessary. This
allows for a reduction in state.

Additionally the logic to build up the option-specific args is now done
prior to merging wiht the sandbox specific args.

Signed-off-by: Caleb Brown <calebbrown@google.com>