Make workers more ephemeral to avoid space issues. (#742)
* Make workers more ephemeral to avoid space issues.

Previously, prior to #172, sandbox images were cached inside the
analysis image.

PR #172 solved this by using a StatefulSet to provide persistent
volumes. However, it didn't properly solve
https://github.com/ossf/package-analysis/pull/172/files#r755697003
where past images continued to consume the space.

Now we are getting errors persistently on nodes where it is likely that
this space is fully occupied.

This change removes the persistent volume, for ephemeral storage. The
emptydir ensures that data presists between crashes, but can be purged
by restarting the pod.

Signed-off-by: Caleb Brown <calebbrown@google.com>

* Further bump replicas

Signed-off-by: Caleb Brown <calebbrown@google.com>

* Bump base replicas

Signed-off-by: Caleb Brown <calebbrown@google.com>

---------

Signed-off-by: Caleb Brown <calebbrown@google.com>